Dolphin OCR/requirements.txt
beautifulsoup4==4.13.4
pdf2image==1.17.0
python_docx==1.1.2
streamlit==1.45.1

albumentations==1.4.0
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
pillow==9.3.0
timm==0.5.4
torch==2.1.0
torchvision==0.16.0
transformers==4.47.0
accelerate==1.6.0
pymupdf==1.26